requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
pandas
openai
chromadb
python-dotenv
tiktoken